I have created a folder called AvraamContribution. I have created the folder structure that will help us keep the project clean. I create some abstract classes such as Weapon that other types of weapon will inherit from and Enemy that other types of enemies will inherit from.
I know that this might sound a bit advanced but if you allow me I can help you int his and have a great and clean projects that artists can collaborate easily with you and add more enemies and weapons.

The scene Cleanup Arcade is not ready to be used but if we have a quick call I can explain my logic and you can pick up the pace from there.

Best, Avraam.